+++ nara/__init__.py
@@ -1,0 +1,13 @@
+import collections
+import os
+import pdb
+try:
+    import pdbp  # (Pdb+) --- Python Debugger Plus
+except Exception:
+    pass
+import sys
+
+from nara.nara.GenFunc import CreateFunc
+from nara.nara.GenTemplate import CreateTemplate
+from nara.nara.init import init
+from nara.nara.tele_cloude_storage import tele_db