+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 0.2 - 13-06-2024
+* Import changes <s>`from Nara.Extra import MailUrl`</s> -> `from nara.extra import MailUrl`
+* `Tele Clude Storage` new feature
+* `tempmailio` new feature
+* `Current Date Time of any country` new feature
+* `sql mannager` new feature
+* Update Docs
+
+
+
 ### 0.1.5 - 20-05-2024
 
 * Fixed API bug